@@ -31,7 +31,7 @@
     "sinon": "^7.2.7"
   },
   "engines": {
-    "node": "8"
+    "node": "11"
   },
   "jest": {
     "testEnvironment": "node"
